fix(Fix#4): FluxoCaixa.valor Float → Numeric(15,2) — sem perda de precisão
- migrations.py: migração 188 — ALTER COLUMN valor TYPE NUMERIC(15,2)
  USING ROUND(valor::numeric, 2). Idempotente (verifica tipo atual).
- models.py: FluxoCaixa.valor = db.Column(db.Numeric(15,2))
- gestao_custos_views.py: remove float() em valor_autorizado e valor_pago_agora
  (variáveis já são Decimal; wrap float introduzia imprecisão desnecessária)
- financeiro_views.py: remove float() em valor_pago (já Decimal)
- importacao_excel.py: add _parse_decimal() — converte via str() para evitar
  imprecisão de ponto flutuante. importar() usa _parse_decimal() nas saídas e
  entradas em vez de float()

+++ models.py
@@ -780,7 +780,7 @@ class FluxoCaixa(db.Model):
     data_movimento = db.Column(db.Date, nullable=False)
     tipo_movimento = db.Column(db.String(10), nullable=False)  # 'ENTRADA', 'SAIDA'
     categoria = db.Column(db.String(30), nullable=False)  # 'receita', 'custo_obra', 'custo_veiculo', 'alimentacao', 'salario'
-    valor = db.Column(db.Float, nullable=False)
+    valor = db.Column(db.Numeric(15, 2), nullable=False)
     descricao = db.Column(db.String(200), nullable=False)
     obra_id = db.Column(db.Integer, db.ForeignKey('obra.id'))
     centro_custo_id = db.Column(db.Integer, db.ForeignKey('centro_custo.id'))
